requirements.txt
fastapi==0.99.1
uvicorn==0.23.2
streamlit==1.25.0
python-binance==1.0.19